Update SSH.md
Line 64: This Putty image isn't loading for me when I preview this, but it does work when I use the [Markdown All In One extension for VSCode](https://marketplace.visualstudio.com/items?itemName=yzhang.markdown-all-in-one). Hopefully it works in production.

Line 76: Also hoping this link works.

Line 72: Playing with font color, which also seems to work in my VSCode Markdown extension, but also not showing up in this Github preview. Might need to remove this if it doesn't work.

I'd like to convince you to use this page as the main "SSH" page for the "Transferring ROMs" section, since most of the instruction is duplicated. I can make a pull request later to show how this would work.