Update Transferring-Roms.md
Proposed edit for the Transferring ROMs page, moving the SFTP/SSH instruction to the SSH page. 

Also, I've been tweaking the headers so that top-level headers have one # and sub-headers have two ##, but I'm beginning to wonder if this is a mistake after looking into the Mkdocs header baselevels. Should we start with headers at two ##? If so, this would be good to add to a "How to Contribute" section.