Update First-Installation.md
## Installation

* If the number of raspberries shown at boot is 4, it could mean a Pi 2, 3, or 4, of which there are two different downloads for, which makes the graphic somewhat useless/confusing. I'm replacing this with some more concrete methods of discovering Pi version.
* While writing the instructions for the SD card reader, I realized it's really better off in the "Required" section, because the user needs *some* way of plugging it into their PC, so I fixed up the Hardware section a bit.

I was going to be gentle with this Installation section and maybe have a new "Quick Setup" headered section right after Hardware, but then I thought... fuck it. I'll just rewrite the whole thing. 

I'm leaning heavily on the Raspberry Pi Imager as discussed in previous PRs. If it does all the checksums and it's an official RPI program and it's mostly future-proof, then this is literally all that's needed. We don't need to download the image separately, then extract it, etc etc. Just use the damn program. Done.

Because this program is the main focus of first-time setups, its instruction needs to be solid. I opted to write out the instructions *and* include a short gif showing how the program is used for those of us that really preferred pop-up books growing up and have no patience for long videos or walls of text.

The advantage is, aside from the above "Video Cable" list which may or may not change in the near future (maybe DisplayPort for the RPI 5?), this entire page is now completely future-proof and barely needs to be touched after this merge. Most everything that can change is linked outside of this page.